bo5: decompile e_dai_no2_shortcut via shared header
