+++ config/splat.pspeu.bobo5.yaml
@@ -69,7 +69,7 @@ segments:
       - [0x16270, c, bo5_psp/unk_16270]
       - [0x18730, c, bo5/e_lock_camera]
       - [0x18990, c, bo5_psp/e_rbo3door]
-      - [0x18E30, c, bo5_psp/e_dai_no2_shortcut]
+      - [0x18E30, c, bo5/e_dai_no2_shortcut]
       - [0x191F0, c, bo5_psp/e_stained_glass]
       - [0x1AD00, c, bo5_psp/e_clouds]
       - [0x1B480, .data, bo5/e_life_up]
@@ -87,7 +87,8 @@ segments:
       - [0x36150, .data, bo5/header]
       - [0x36180, data]
       - [0x36AE8, .data, bo5/e_lock_camera]
-      - [0x36B60, data, 36B60]
+      - [0x36B60, .data, bo5/e_dai_no2_shortcut]
+      - [0x36B68, data, 36B68]
       - [0x36F00, .rodata, bo5/e_life_up]
       - [0x36F20, .rodata, bo5/e_red_door]
       - [0x36F38, .rodata, bo5/e_collect]
