+++ src/boss/bo5/e_dai_no2_shortcut.c
@@ -1,6 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 #include "bo5.h"
 
-INCLUDE_ASM("boss/bo5/nonmatchings/e_dai_no2_shortcut", EntityBlock);
-
-INCLUDE_ASM("boss/bo5/nonmatchings/e_dai_no2_shortcut", EntityStatue);
+#include "../../st/e_dai_no2_shortcut.h"
