--- src/boss/bo5_psp/e_dai_no2_shortcut.c
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../bo5/bo5.h"
-
-INCLUDE_ASM("boss/bo5_psp/nonmatchings/bo5_psp/e_dai_no2_shortcut", EntityBlock);
-
-INCLUDE_ASM("boss/bo5_psp/nonmatchings/bo5_psp/e_dai_no2_shortcut", EntityStatue);
